Add argument --database-do-not-migrate.
