+++ spec/lib/hets/logic_graph_request_spec.rb
@@ -9,32 +9,16 @@ describe HetsAgent::Hets::LogicGraphRequest do
 
   subject { HetsAgent::Hets::LogicGraphRequest.new }
 
-  context 'request' do
-    it_behaves_like 'a HetsAgent::Hets::Request'
-  end
+  it_behaves_like 'a HetsAgent::Hets::Request'
+  it_behaves_like 'a database request'
 
   context 'arguments' do
     it 'logic graph' do
       expect(subject.arguments).to include('--logic-graph')
     end
 
-    it 'output type' do
-      expect(subject.arguments).to include('--output-types=db')
-    end
-
-    it 'database.yml' do
-      database_yml = HetsAgent::Application.root.join('config/database.yml')
-      expect(subject.arguments).
-        to include("--database-config=#{database_yml}")
-    end
-
-    it 'database subconfig' do
-      expect(subject.arguments).
-        to include("--database-subconfig=#{HetsAgent::Application.env}")
-    end
-
     it 'do not contain anything else' do
-      expect(subject.arguments.length).to eq(5)
+      expect(subject.arguments.length).to eq(6)
     end
   end
 end
